# apps/admin-web/src/api/generated/models/templateVersionRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * 万企商服通接口文档
 * 面向万企商服通小程序和管理后台的统一接口契约。除明确标记为公开的接口外，均需携带 Bearer Token。
 * OpenAPI spec version: 1.0.0
 */
import type { TemplateItemRequest } from './templateItemRequest';

export interface TemplateVersionRequest {
  /** 接口字段 */
  templateName: string;
  /** 当前页数据 */
  items: TemplateItemRequest[];
}