# apps/client-uni/src/api/generated/models/apiResponseTemplateView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * 万企商服通接口文档
 * 面向万企商服通小程序和管理后台的统一接口契约。除明确标记为公开的接口外，均需携带 Bearer Token。
 * OpenAPI spec version: 1.0.0
 */
import type { TemplateView } from './templateView';

export interface ApiResponseTemplateView {
  /** 请求是否成功 */
  success?: boolean;
  /** 业务结果编码 */
  code?: string;
  /** 中文结果说明 */
  message?: string;
  /** 业务数据 */
  data?: TemplateView;
  /** 请求追踪编号 */
  requestId?: string;
  /** 响应时间 */
  timestamp?: string;
}