# openapi/万企商服通_OpenAPI_V1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "万企商服通接口文档",
    "description": "面向万企商服通小程序和管理后台的统一接口契约。除明确标记为公开的接口外，均需携带 Bearer Token。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前服务地址"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/templates/{templateId}/publish": {
      "post": {
        "tags": [
          "后台产品"
        ],
        "summary": "发布资料模板版本",
        "description": "发布资料模板版本。响应统一使用 ApiResponse 结构。 ",
        "operationId": "adminProductPublishTemplate",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "description": "模板编号",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTemplateView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/templates/{sourceTemplateId}/versions": {
      "post": {
        "tags": [
          "后台产品"
        ],
        "summary": "创建资料模板版本",
        "description": "创建资料模板版本。响应统一使用 ApiResponse 结构。 ",
        "operationId": "adminProductCreateVersion",
        "parameters": [
          {
            "name": "sourceTemplateId",
            "in": "path",
            "description": "来源模板编号",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "请求参数",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTemplateView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/products": {
      "get": {
        "tags": [
          "后台产品"
        ],
        "summary": "查询后台产品列表",
        "description": "查询后台产品列表。响应统一使用 ApiResponse 结构。 ",
        "operationId": "adminProductList",
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListProductAdminView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "后台产品"
        ],
        "summary": "创建服务产品",
        "description": "创建服务产品。响应统一使用 ApiResponse 结构。 ",
        "operationId": "adminProductCreate",
        "requestBody": {
          "description": "请求参数",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductAdminView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/products/{productId}/unpublish": {
      "post": {
        "tags": [
          "后台产品"
        ],
        "summary": "下架服务产品",
        "description": "下架服务产品。响应统一使用 ApiResponse 结构。 ",
        "operationId": "adminProductUnpublish",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "产品编号",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductAdminView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/products/{productId}/publish": {
      "post": {
        "tags": [
          "后台产品"
        ],
        "summary": "发布服务产品",
        "description": "发布服务产品。响应统一使用 ApiResponse 结构。 ",
        "operationId": "adminProductPublish",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "产品编号",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductAdminView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/imports": {
      "post": {
        "tags": [
          "后台导入"
        ],
        "summary": "创建企业 Excel 导入任务",
        "description": "创建企业 Excel 导入任务。响应统一使用 ApiResponse 结构。 ",
        "operationId": "importCreate",
        "parameters": [
          {
            "name": "dryRun",
            "in": "query",
            "description": "是否仅校验不写入",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "接口参数",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "请求参数",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseJobView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/imports/{jobId}/error-report": {
      "post": {
        "tags": [
          "后台导入"
        ],
        "summary": "获取导入错误报告",
        "description": "获取导入错误报告。响应统一使用 ApiResponse 结构。 ",
        "operationId": "importError",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "任务编号",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTemplateDownload"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/products/{productId}": {
      "patch": {
        "tags": [
          "后台产品"
        ],
        "summary": "修改服务产品",
        "description": "修改服务产品。响应统一使用 ApiResponse 结构。 ",
        "operationId": "adminProductUpdate",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "产品编号",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "请求参数",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductAdminView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/imports/{jobId}": {
      "get": {
        "tags": [
          "后台导入"
        ],
        "summary": "查询企业导入任务",
        "description": "查询企业导入任务。响应统一使用 ApiResponse 结构。 ",
        "operationId": "importGet",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "任务编号",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseJobView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseTemplateView": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "code": {
            "type": "string",
            "description": "业务结果编码"
          },
          "message": {
            "type": "string",
            "description": "中文结果说明"
          },
          "data": {
            "$ref": "#/components/schemas/TemplateView",
            "description": "业务数据"
          },
          "requestId": {
            "type": "string",
            "description": "请求追踪编号"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "响应时间"
          }
        }
      },
      "ItemView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键编号"
          },
          "itemCode": {
            "type": "string",
            "description": "接口字段"
          },
          "itemName": {
            "type": "string",
            "description": "接口字段"
          },
          "inputType": {
            "type": "string",
            "description": "录入类型"
          },
          "required": {
            "type": "boolean",
            "description": "是否必填"
          },
          "sensitive": {
            "type": "boolean",
            "description": "是否敏感"
          },
          "allowedExtensions": {
            "type": "array",
            "description": "接口字段",
            "items": {
              "type": "string"
            }
          },
          "maxFileSizeMb": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "maxFileCount": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "helpText": {
            "type": "string",
            "description": "接口字段"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          }
        }
      },
      "TemplateView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键编号"
          },
          "templateCode": {
            "type": "string",
            "description": "接口字段"
          },
          "templateName": {
            "type": "string",
            "description": "接口字段"
          },
          "versionNo": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "items": {
            "type": "array",
            "description": "当前页数据",
            "items": {
              "$ref": "#/components/schemas/ItemView"
            }
          }
        }
      },
      "TemplateItemRequest": {
        "type": "object",
        "properties": {
          "itemCode": {
            "type": "string",
            "description": "接口字段"
          },
          "itemName": {
            "type": "string",
            "description": "接口字段"
          },
          "inputType": {
            "type": "string",
            "description": "录入类型"
          },
          "required": {
            "type": "boolean",
            "description": "是否必填"
          },
          "sensitive": {
            "type": "boolean",
            "description": "是否敏感"
          },
          "allowedExtensions": {
            "type": "string",
            "description": "接口字段"
          },
          "maxFileSizeMb": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "maxFileCount": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "helpText": {
            "type": "string",
            "description": "接口字段"
          }
        },
        "required": [
          "inputType",
          "itemCode",
          "itemName",
          "maxFileCount"
        ]
      },
      "TemplateVersionRequest": {
        "type": "object",
        "properties": {
          "templateName": {
            "type": "string",
            "description": "接口字段"
          },
          "items": {
            "type": "array",
            "description": "当前页数据",
            "items": {
              "$ref": "#/components/schemas/TemplateItemRequest"
            }
          }
        },
        "required": [
          "items",
          "templateName"
        ]
      },
      "ProductRequest": {
        "type": "object",
        "properties": {
          "moduleId": {
            "type": "integer",
            "format": "int64",
            "description": "接口字段"
          },
          "productCode": {
            "type": "string",
            "description": "接口字段"
          },
          "productName": {
            "type": "string",
            "description": "接口字段"
          },
          "description": {
            "type": "string",
            "description": "接口字段"
          },
          "templateId": {
            "type": "integer",
            "format": "int64",
            "description": "模板编号"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "description": "乐观锁版本号"
          }
        },
        "required": [
          "moduleId",
          "productCode",
          "productName",
          "sortOrder",
          "templateId"
        ]
      },
      "ApiResponseProductAdminView": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "code": {
            "type": "string",
            "description": "业务结果编码"
          },
          "message": {
            "type": "string",
            "description": "中文结果说明"
          },
          "data": {
            "$ref": "#/components/schemas/ProductAdminView",
            "description": "业务数据"
          },
          "requestId": {
            "type": "string",
            "description": "请求追踪编号"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "响应时间"
          }
        }
      },
      "ProductAdminView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键编号"
          },
          "moduleId": {
            "type": "integer",
            "format": "int64",
            "description": "接口字段"
          },
          "productCode": {
            "type": "string",
            "description": "接口字段"
          },
          "productName": {
            "type": "string",
            "description": "接口字段"
          },
          "description": {
            "type": "string",
            "description": "接口字段"
          },
          "templateId": {
            "type": "integer",
            "format": "int64",
            "description": "模板编号"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "status": {
            "type": "string",
            "description": "业务状态"
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "description": "乐观锁版本号"
          }
        }
      },
...
      "ApiResponseJobView": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "code": {
            "type": "string",
            "description": "业务结果编码"
          },
          "message": {
            "type": "string",
            "description": "中文结果说明"
          },
          "data": {
            "$ref": "#/components/schemas/JobView",
            "description": "业务数据"
          },
          "requestId": {
            "type": "string",
            "description": "请求追踪编号"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "响应时间"
          }
        }
      },
      "JobView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键编号"
          },
          "jobNo": {
            "type": "string",
            "description": "接口字段"
          },
          "status": {
            "type": "string",
            "description": "业务状态"
          },
          "dryRun": {
            "type": "boolean",
            "description": "是否仅校验不写入"
          },
          "totalRows": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "successRows": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "skippedRows": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "failedRows": {
            "type": "integer",
            "format": "int32",
            "description": "接口字段"
          },
          "errorReportFileId": {
            "type": "integer",
            "format": "int64",
            "description": "接口字段"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "description": "完成时间"
          }
        }
      },
      "ApiResponseTemplateDownload": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "code": {
            "type": "string",
            "description": "业务结果编码"
          },
          "message": {
            "type": "string",
            "description": "中文结果说明"
          },
          "data": {
            "$ref": "#/components/schemas/TemplateDownload",
            "description": "业务数据"
          },
          "requestId": {
            "type": "string",
            "description": "请求追踪编号"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "响应时间"
          }
        }
      },
      "TemplateDownload": {
        "type": "object",
        "properties": {
          "downloadUrl": {
            "type": "string",
            "description": "短时有效下载地址"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "失效时间"
          }
        }
      },
...
      "ApiResponseListProductAdminView": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "code": {
            "type": "string",
            "description": "业务结果编码"
          },
          "message": {
            "type": "string",
            "description": "中文结果说明"
          },
          "data": {
            "type": "array",
            "description": "业务数据",
            "items": {
              "$ref": "#/components/schemas/ProductAdminView"
            }
          },
          "requestId": {
            "type": "string",
            "description": "请求追踪编号"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "响应时间"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "登录接口返回的访问令牌",
        "scheme": "bearer",
        "bearerFormat": "Sa-Token"
      }
    }
  }
}